Extract printBlock
